Pickles: initial verification for custom gates
Fixes #12486.

- Expose custom gates commitments through kimchi_bindings
- Add support within Pickles_types for these gates
- Propagate flags throughout the code (in-circuit / out-of-circuit)
- Expand `Common.ft_comm` to use the incoming commitments

+++ src/lib/crypto/kimchi_bindings/stubs/src/pasta_fp_plonk_proof.rs
@@ -122,12 +122,12 @@ pub fn caml_pasta_fp_plonk_proof_example_with_lookup(
     for table_id in 0..num_tables {
         let cfg = if indexed {
             RuntimeTableCfg::Indexed(RuntimeTableSpec {
-                id: table_id as i32,
+                id: table_id,
                 len: 5,
             })
         } else {
             RuntimeTableCfg::Custom {
-                id: table_id as i32,
+                id: table_id,
                 first_column: [8u32, 9, 8, 7, 1].into_iter().map(Into::into).collect(),
             }
         };
